+++ src/components/dropdown-option/script.ts
@@ -1,0 +1,17 @@
+import {
+  Component,
+  Prop,
+  Vue,
+} from 'vue-property-decorator';
+
+
+@Component({
+  name: 'StDropdownOption',
+})
+export default class StDropdownOption extends Vue {
+  @Prop(Boolean)
+  readonly!: boolean;
+
+  @Prop(Boolean)
+  disabled!: boolean;
+}